Telegram Bot
A bot for Telegram (messaging app). The user sends the bot a message, and the bot “reads” the user’s text aloud by sending them a text-to-speech audio file of their text. Additionally, the bot interacts with a user’s Raspberry Pi, turning an LED on when instructed to over text.

This is the JSON file for it.